+++ android/src/main/java/com/lib/flutter_blue_plus/FlutterBluePlusPlugin.java
@@ -282,7 +282,7 @@ public class FlutterBluePlusPlugin implements
             // check that we have an adapter, except for 
             // the functions that do not need it
             if(mBluetoothAdapter == null && 
-                "flutterHotRestart".equals(call.method) == false &&
+                "flutterRestart".equals(call.method) == false &&
                 "connectedCount".equals(call.method) == false &&
                 "setLogLevel".equals(call.method) == false &&
                 "isSupported".equals(call.method) == false &&
@@ -294,7 +294,7 @@ public class FlutterBluePlusPlugin implements
 
             switch (call.method) {
 
-                case "flutterHotRestart":
+                case "flutterRestart":
                 {
                     // no adapter?
                     if (mBluetoothAdapter == null) {
@@ -309,7 +309,9 @@ public class FlutterBluePlusPlugin implements
                         mIsScanning = false;
                     }
 
-                    disconnectAllDevices("flutterHotRestart");
+                    // all dart state is reset after flutter restart
+                    // (i.e. Hot Restart) so also reset native state
+                    disconnectAllDevices("flutterRestart");
 
                     log(LogLevel.DEBUG, "connectedPeripherals: " + mConnectedDevices.size());
 
@@ -733,7 +735,7 @@ public class FlutterBluePlusPlugin implements
                         }
                     }
                     if (gatt == null) {
-                        gatt = mConnectedDevices.get(remoteId);;
+                        gatt = mConnectedDevices.get(remoteId);
                     }
                     if (gatt == null) {
                         gatt = mAutoConnected.get(remoteId);
@@ -768,11 +770,14 @@ public class FlutterBluePlusPlugin implements
                         // cleanup
                         gatt.close();
 
+                        // random number defined by flutter blue plus.
+                        int bmUserCanceledErrorCode = 23789258;
+
                         // see: BmConnectionStateResponse
                         HashMap<String, Object> response = new HashMap<>();
                         response.put("remote_id", remoteId);
                         response.put("connection_state", bmConnectionStateEnum(BluetoothProfile.STATE_DISCONNECTED));
-                        response.put("disconnect_reason_code", 23789258); // random value
+                        response.put("disconnect_reason_code", bmUserCanceledErrorCode);
                         response.put("disconnect_reason_string", "connection canceled");
 
                         invokeMethodUIThread("OnConnectionStateChanged", response);
@@ -1753,20 +1758,19 @@ public class FlutterBluePlusPlugin implements
         int n = 0;
 
         while (n < bytes.length) {
-
-            int fieldLen = bytes[n];
+            int fieldLen = bytes[n] & 0xFF;
 
             // no more or malformed data
             if (fieldLen <= 0) {
                 break;
             }
 
-            // end of packet
-            if (fieldLen + n > bytes.length - 1) {
+            // Ensuring we don't go past the bytes array
+            if (n + fieldLen >= bytes.length) {
                 break;
             }
 
-            int dataType = bytes[n + 1];
+            int dataType = bytes[n + 1] & 0xFF;
 
             // no more data
             if (dataType == 0) {
@@ -1775,15 +1779,82 @@ public class FlutterBluePlusPlugin implements
 
             // appearance type byte
             if (dataType == 0x19 && fieldLen == 3) {
-                int loByte = bytes[n + 2] & 0xFF;
-                int hiByte = bytes[n + 3] & 0xFF;
-                return hiByte * 256 + loByte;
+                int high = (bytes[n + 3] & 0xFF) << 8;
+                int low = (bytes[n + 2] & 0xFF);
+                return high | low;
             }
 
             n += fieldLen + 1;
         }
 
         return 0;
+    }
+
+    Map<Integer, byte[]> getManufacturerSpecificData(ScanRecord adv) {
+        byte[] bytes = adv.getBytes();
+        Map<Integer, byte[]> manufacturerDataMap = new HashMap<>();
+        int n = 0;
+        while (n < bytes.length) {
+
+            // layout:
+            // n[0] = fieldlen
+            // n[1] = datatype (MSD)
+            // n[2] = manufacturerId (low)
+            // n[3] = manufacturerId (high)
+            // n[4] = data...
+            int fieldLen = bytes[n] & 0xFF;
+
+            // no more or malformed data
+            if (fieldLen <= 0) {
+                break;
+            }
+
+            // Ensuring we don't go past the bytes array
+            if (n + fieldLen >= bytes.length) {
+                break;
+            }
+
+            int dataType = bytes[n + 1] & 0xFF;
+
+            // Manufacturer Specific Data magic number
+            // At least 3 bytes: 2 for manufacturer ID & 1 for dataType
+            if (dataType == 0xFF && fieldLen >= 3) {
+
+                // Manufacturer Id
+                int high = (bytes[n + 3] & 0xFF) << 8;
+                int low = (bytes[n + 2] & 0xFF);
+                int manufacturerId = high | low;
+
+                // the length of the msd data,
+                // excluding manufacturerId & dataType
+                int msdLen = fieldLen - 3;
+
+                // ptr to msd data
+                // excluding manufacturerId & dataType
+                int msdPtr = n + 4;
+
+                // add to map
+                if (manufacturerDataMap.containsKey(manufacturerId)) {
+                    // If the manufacturer ID already exists, append the new data to the existing list
+                    byte[] existingData = manufacturerDataMap.get(manufacturerId);
+                    byte[] mergedData = new byte[existingData.length + msdLen];
+                    // Merge arrays
+                    System.arraycopy(existingData, 0, mergedData, 0, existingData.length);
+                    System.arraycopy(bytes, msdPtr, mergedData, existingData.length, msdLen);
+                    manufacturerDataMap.put(manufacturerId, mergedData);
+                } else {
+                    // Otherwise, put the new manufacturer ID and its data into the map
+                    byte[] data = new byte[msdLen];
+                    // Starting from n+4 because manufacturerId occupies n+2 and n+3
+                    System.arraycopy(bytes, msdPtr, data, 0, data.length);
+                    manufacturerDataMap.put(manufacturerId, data);
+                }
+            }
+
+            n += fieldLen + 1;
+        }
+
+        return manufacturerDataMap;
     }
 
     /////////////////////////////////////////////////////////////////////////////////////
@@ -1814,6 +1885,19 @@ public class FlutterBluePlusPlugin implements
             final int adapterState = intent.getIntExtra(BluetoothAdapter.EXTRA_STATE, BluetoothAdapter.ERROR);
 
             log(LogLevel.DEBUG, "OnAdapterStateChanged: " + adapterStateString(adapterState));
+
+            // stop scanning when adapter is turned off. 
+            // Otherwise, scanning automatically resumes when the adapter is
+            // turned back on. I don't think most users expect that.
+            if (adapterState != BluetoothAdapter.STATE_ON) {
+                if (mBluetoothAdapter != null && mIsScanning) {
+                    BluetoothLeScanner scanner = mBluetoothAdapter.getBluetoothLeScanner();
+                    if (scanner != null) {
+                        scanner.stopScan(getScanCallback());
+                        mIsScanning = false;
+                    }
+                }
+            }
             
             // see: BmBluetoothAdapterState
             HashMap<String, Object> map = new HashMap<>();
@@ -2356,7 +2440,7 @@ public class FlutterBluePlusPlugin implements
         ScanRecord adv = result.getScanRecord();
 
         boolean connectable;
-        if(Build.VERSION.SDK_INT >= 26) { // Android 8.0, August 2017
+        if (Build.VERSION.SDK_INT >= 26) { // Android 8.0, August 2017
             connectable = result.isConnectable();
         } else {
             // Prior to Android 8.0, it is not possible to get if connectable.
@@ -2368,23 +2452,21 @@ public class FlutterBluePlusPlugin implements
         String                  advName      = adv != null ?  adv.getDeviceName()                : null;
         int                     txPower      = adv != null ?  adv.getTxPowerLevel()              : min;
         int                     appearance   = adv != null ?  getAppearanceFromScanRecord(adv)   : 0;
-        SparseArray<byte[]>     manufData    = adv != null ?  adv.getManufacturerSpecificData()  : null;
+        Map<Integer, byte[]>    manufData    = adv != null ?  getManufacturerSpecificData(adv)   : null;
         List<ParcelUuid>        serviceUuids = adv != null ?  adv.getServiceUuids()              : null;
         Map<ParcelUuid, byte[]> serviceData  = adv != null ?  adv.getServiceData()               : null;
 
         // Manufacturer Specific Data
-        HashMap<Integer, String> manufDataB = new HashMap<Integer, String>();
-        if(manufData != null) {
-            for (int i = 0; i < manufData.size(); i++) {
-                int key = manufData.keyAt(i);
-                byte[] value = manufData.valueAt(i);
-                manufDataB.put(key, bytesToHex(value));
+        HashMap<Integer, String> manufDataB = new HashMap<>();
+        if (manufData != null) {
+            for (Map.Entry<Integer, byte[]> entry : manufData.entrySet()) {
+                manufDataB.put(entry.getKey(), bytesToHex(entry.getValue()));
             }
         }
 
         // Service Data
         HashMap<String, Object> serviceDataB = new HashMap<>();
-        if(serviceData != null) {
+        if (serviceData != null) {
             for (Map.Entry<ParcelUuid, byte[]> entry : serviceData.entrySet()) {
                 ParcelUuid key = entry.getKey();
                 byte[] value = entry.getValue();
@@ -2393,8 +2475,8 @@ public class FlutterBluePlusPlugin implements
         }
 
         // Service UUIDs
-        List<String> serviceUuidsB = new ArrayList<String>();
-        if(serviceUuids != null) {
+        List<String> serviceUuidsB = new ArrayList<>();
+        if (serviceUuids != null) {
             for (ParcelUuid s : serviceUuids) {
                 serviceUuidsB.add(uuidStr(s.getUuid()));
             }